[IMP] manifest loading in steps
To lower the rebuilds needed when a change is made in data files, the module initialization has been split into:
- symbol creation: we load the manifest content (needed to sort module loading)
- Arch: normal arch python on `__init__.py`
- Arch_eval: load data files (was previously in symbol creation). Load them only if symbol does not already exists.
- Validation: python and manifest validation.
