+++ server/src/core/csv_validation.rs
@@ -2,17 +2,14 @@ use std::path::PathBuf;
 
 use csv::StringRecord;
 use lsp_types::{Diagnostic, Position, Range};
+use tracing::info;
 
 use crate::{
-    constants::{BuildStatus, BuildSteps, OYarn},
-    core::{
-        diagnostics::{create_diagnostic, DiagnosticCode},
+    Sy, constants::{BuildStatus, BuildSteps, DEBUG_STEPS, OYarn}, core::{
+        diagnostics::{DiagnosticCode, create_diagnostic},
         file_mgr::FileInfo,
-        symbols::{symbol_keys::{CsvFileKey, ModuleKey}, storage::SymbolTable},
-    },
-    features::csv_ast_utils::CsvFieldIter,
-    threads::SessionInfo,
-    Sy,
+        symbols::{storage::SymbolTable, symbol_keys::{CsvFileKey, ModuleKey}},
+    }, features::csv_ast_utils::CsvFieldIter, threads::SessionInfo
 };
 use std::{cell::RefCell, rc::Rc};
 
@@ -30,6 +27,9 @@ impl CsvValidator {
         let mut diagnostics = vec![];
         session.st_mut().set_build_status(csv_symbol.into(), BuildSteps::VALIDATION, BuildStatus::IN_PROGRESS);
         let path = session.st()[csv_symbol].path.clone();
+        if DEBUG_STEPS {
+            info!("VALIDATION - CSV: {}", path);
+        }
         let Some(file_info) = SymbolTable::get_file_info_for_validation(session, csv_symbol.into()) else {
             session.st_mut().set_build_status(csv_symbol.into(), BuildSteps::VALIDATION, BuildStatus::INVALID);
             return;
